\name{chemopac-package}
\alias{chemopac-package}
\alias{chemopac}
\docType{package}
\title{
\packageTitle{chemopac}
}
\description{
\packageDescription{chemopac}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{chemopac}
\packageIndices{chemopac}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{chemopac}

Maintainer: \packageMaintainer{chemopac}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
